+++ DSAAProject/src/DataAnalysis/BornPlace.java
@@ -34,9 +34,7 @@ public class BornPlace {
     public static void main(String[] args) {
         CSVReader reader = new CSVReader("DSAAProject\\FileStorage\\Project_data_20181208.csv"); // Use the relative path of the .csv file
         ArrayList<Student> students = reader.parse(); // Format the data into a list
-
     }
-
 
     private HashMap<String, Province> province;
     private MyTreeMap<String, Province> sortedProvince;
@@ -64,8 +62,6 @@ public class BornPlace {
         return s.toString();
 
     }
-
-
 }
 
 class Province extends ComparableNode<Province> implements AddAble<Student> {
@@ -78,6 +74,7 @@ class Province extends ComparableNode<Province> implements AddAble<Student> {
      *
      * @param provinceName is a String
      */
+
     Province(String provinceName, Student s) {
         this.provinceName = provinceName;
         num = 1;// This is in the super class, stands for the Number of City
@@ -105,7 +102,6 @@ class Province extends ComparableNode<Province> implements AddAble<Student> {
             City c = city.remove(name);
             c.add(s);
             city.put(name, c);//This is because get means remove
-
         } else {
             city.put(name, new City(name, s));
         }
@@ -121,7 +117,6 @@ class Province extends ComparableNode<Province> implements AddAble<Student> {
         return "--" + this.provinceName + "(" + num + ")" + "\n" + s;
     }
 }
-
 
 class City extends ComparableNode<City> implements AddAble<Student> {
     String cityName;
@@ -159,7 +154,6 @@ class City extends ComparableNode<City> implements AddAble<Student> {
         for (District d : sortedDistrict.toArray(new District[0])) {
             s.append(d.toString());
         }
-
         return "------" + this.cityName + "(" + num + ")" + "\n" + s;
     }
 }
