+++ DSAAProject/src/Start2.java
@@ -1,4 +1,4 @@
-import DataAnalysis.CountryUni;
+import DataAnalysis.CountryUniMajor;
 import Student.CSVReader;
 import Student.Student;
 
@@ -9,7 +9,7 @@ public class Start2 {
         CSVReader reader = new CSVReader("DSAAProject\\FileStorage\\Project_data_20181208.csv"); // Use the relative path of the .csv file
         ArrayList<Student> students = reader.parse(); // Format the data into a list
 
-        CountryUni countryUni = new CountryUni(students.toArray(new Student[0]));
-        FileIO.fileWriter(countryUni.toString(), "DSAAProject\\FileStorage\\Country Uni.txt");
+        CountryUniMajor countryUniMajor = new CountryUniMajor(students.toArray(new Student[0]));
+        FileIO.fileWriter(countryUniMajor.toString(), "DSAAProject\\FileStorage\\Country Uni Major.txt");
     }
 }
